Refactor bump version 0.2.9
