+++ src/components/Receiver/ReceiverContext.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const ReceiverContext = React.createContext({
+  setPeerAddress: (item: string) => {},
+  toggle: () => {},
+});
+
+export default ReceiverContext;
