+++ src/components/Receiver/index.ts
@@ -1,1 +1,2 @@
-export { default } from "./Receiver";
+export { default as Receiver } from "./Receiver";
+export { default as ReceiverLaunch } from "./ReceiverLaunch";
