--- src/components/ReceiverConfig/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ReceiverConfig";
-export { ReceiverContext } from "./ReceiverConfig";
