+++ src/components/index.ts
@@ -1,2 +1,1 @@
-export { default as Receiver } from "./Receiver";
-export { default as ReceiverConfig } from "./ReceiverConfig";
+export { Receiver, ReceiverLaunch } from "./Receiver";
